+++ hr_employee_firstname/README.rst
@@ -7,7 +7,7 @@ HR Employee First Name, Last Name
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:482b99d666b8bafef983166ebc8c36bfa563076f54886db9eb8a66b6befc8881
+   !! source digest: sha256:ffa31618c18bdd243ee469b30e00a10c4adb1625eb8e0723c119219deaa7bab8
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -85,6 +85,17 @@ This module is maintained by the OCA.
 OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
+
+.. |maintainer-Savoir-faire Linux| image:: https://github.com/Savoir-faire Linux.png?size=40px
+    :target: https://github.com/Savoir-faire Linux
+    :alt: Savoir-faire Linux
+.. |maintainer-luisg123v| image:: https://github.com/luisg123v.png?size=40px
+    :target: https://github.com/luisg123v
+    :alt: luisg123v
+
+Current `maintainers <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-Savoir-faire Linux| |maintainer-luisg123v| 
 
 This module is part of the `OCA/hr <https://github.com/OCA/hr/tree/16.0/hr_employee_firstname>`_ project on GitHub.
